Moved `isHyperlink` and `postParameters` to lib directory
